refactor: Move <InlinedImageList/> to @chainlit/components
