--- frontend/src/components/atoms/element/inlined/images.tsx
@@ -1,17 +1,0 @@
-import { IImageElement } from '@chainlit/components';
-
-import ImageElement from '../image';
-import { ListWithSize } from './utils';
-
-interface Props {
-  items: IImageElement[];
-}
-
-export default function InlinedImageList({ items }: Props) {
-  return (
-    <ListWithSize<IImageElement>
-      elements={items}
-      renderElement={ImageElement}
-    />
-  );
-}
