+++ libs/components/src/elements/index.ts
@@ -6,3 +6,4 @@ export { ImageElement } from './Image';
 // Inlined
 export { InlinedAudioList } from './InlinedAudioList';
 export { InlinedFileList } from './InlinedFileList';
+export { InlinedImageList } from './InlinedImageList';
